+++ app/components/ExploreSourcesTab.jsx
@@ -1,8 +1,9 @@
 import React, {Component} from 'react'
 import styles from '../styles/main.css';
+import exploreStyles from '../styles/explore.css';
 import {Button, Fab} from 'react-mdc-web';
 import {
-    getMobileSourceNames, getMobileSizeMax, getMobileDetailPath, getColor
+    getMobileSourceNames, getMobileSizeMax, getMobileDetailPath, getColor, getRegionToTitleMap
 } from '../utils/getConfig';
 import ol from 'openlayers';
 
@@ -18,7 +19,7 @@ class ExploreSourcesTab extends Component {
 
     render() {
         //ref: https://github.com/pka/ol3-react-example/blob/master/index.js
-        let tabs = [];
+        let contents = [];
         // Mobile
         if (screen.width <= getMobileSizeMax()) {
             let mobile_sourcenames = getMobileSourceNames().toUpperCase();
@@ -35,7 +36,7 @@ class ExploreSourcesTab extends Component {
             }
             mobile_data.map(data => {
                 let location = (getMobileDetailPath() + data.name.toString() + '/separate/');
-                tabs.push(
+                contents.push(
                     <span key={data.id}>
                         <a key={data.id} href={location}>
                             <Button className={styles.exploreButtonMobile} raised key={data.id} id={data.id}>
@@ -46,22 +47,30 @@ class ExploreSourcesTab extends Component {
                     </span>)
             });
         } else {
-            this.props.data.filter(data => data.properties.type.id === this.props.source.id)
-                .map(data => {
-                    // Convert to Web Mercator Format
-                    let lonLatPoint = [data.geometry.coordinates[0],data.geometry.coordinates[1]];
-                    let webMercatorPoint = ol.proj.fromLonLat(lonLatPoint);
-                    let color = getColor(data.properties.type.id);
-                    tabs.push(<Fab key={data.id} className={styles.exploreButton} style={{backgroundColor: color}}
-                                   onClick={this.clickSensor.bind(this, data.id, data.name, webMercatorPoint)}
-                                   id={data.id} title={data.name}>
-                        <span className={styles.exploreButtonText}>{data.id}</span></Fab>)
-                });
+            let regionToTitleMap = getRegionToTitleMap();
+
+            this.props.regions.map(region => {
+                let source_pills = [];
+                this.props.data.filter(data => data.properties.type.id === this.props.source.id &&
+                   data.properties.region === region ).map(data => {
+                       let lonLatPoint = [data.geometry.coordinates[0],data.geometry.coordinates[1]];
+                       let webMercatorPoint = ol.proj.fromLonLat(lonLatPoint);
+                       let color = getColor(data.properties.type.id);
+                    source_pills.push(<Fab key={data.id} className={exploreStyles.exploreButton} style={{backgroundColor: color}}
+                                        onClick={this.clickSensor.bind(this, data.id, data.name, webMercatorPoint)}
+                                        id={data.id} title={data.name}>
+                           <span className={exploreStyles.exploreButtonText}>{data.id}</span></Fab>)
+                   });
+                if(source_pills.length > 0) {
+                    const title = regionToTitleMap[region];
+                    contents.push(<div key={region}><div><span data-tooltip={title} className={exploreStyles.regionLabel}>{region}</span> </div>{source_pills}</div>)
+                }
+            });
         }
 
         return (
             <div key={this.id}>
-                {tabs}
+                {contents}
             </div>
         );
     }